fix linting rename app temporarily
