+++ libraries/appframeworks/portals/tsconfig.spec.json
@@ -1,6 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
+    "composite": true,
     "outDir": "../../../dist/out-tsc",
     "types": [
       "vitest/globals",
