+++ libraries/mapping/carma-map-layers/src/components/LibModal.tsx
@@ -37,7 +37,7 @@ export interface LibModalProps {
   customCategories?: any[];
 }
 
-const LibModal = ({
+export const LibModal = ({
   open,
   setOpen,
   setAdditionalLayers,
